linger-users: disable linger for zero lingering users
The systemd unit for adding and removing lingering user configuration is
only enabled if there are users configured with lingering in the NixOS
configuration.  This means that if a NixOS system is built with some
lingering users, then the linger configuration is removed from all those
users, the script to disable lingering won't be run, and those users
will incorrectly continue to have lingering enabled.

To avoid this, always build the systemd unit for updating lingering
users, but make running it conditional on there being users configured
to linger in NixOS *or* on there being users configured to linger
according to systemd's loginctl.