tests/systemd-user-linger: test clearing lingering users
Split the existing system-user-linger tests into two: one which tests
lingering is set up for the correct users and only the correct users,
and one which checks that lingering is cleared for users that don't
exist.

This means that the latter test now also ensures the lingering
configuration is handled even if there are no users configured who
should be lingering, where previously such lingering configuration
wouldn't be touched.