git: build documentation in build phase
For build process clarity, build documentation during the build phase.
Without this change, the docs are built during the install phase as
that's when `make` realises they're not yet present.

This was the intent of the change in 2b5899382f6c (git: build documentation
in build phase, 2025-01-03).  However, the Git documentation makefile
appears to have a bug: as of a38edab7c8 (Makefile: generate doc versions
via GIT-VERSION-GEN, 2024-12-06) in the Git repository, the default
build target stopped being "all" and started being a single
documentation config file.